Manually convert all template project to v9
Unfortunately there is no kicad-cli command which allows to automate the
update of our template.
So every single template has to be opened as a project, then schematic
and board need to be individually saved. Same for included symbols and
footprints.
While at it, I also updated symbols and footprints in the respective
templates to v9. Since they mostly consist of pinheaders, there should
not be any real change. But making the ERC and DRC happy for its users.

The only real change is that the edgeberry symbol now has a filled box.